moves model classes of api schemas to models_library
