+++ packages/models-library/src/models_library/api_schemas_webserver/projects_permalink.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel, HttpUrl
+
+
+class ProjectPermalink(BaseModel):
+    url: HttpUrl
+    is_public: bool
